+++ src/main/java/insane96mcp/iguanatweaksreborn/module/items/StackSizes.java
@@ -34,7 +34,7 @@ public class StackSizes extends Feature {
 
     @Config
     @Label(name = "Food Stack Reduction Formula", description = "The formula to calculate the stack size of a food item. Variables as hunger, saturation_modifier, effectiveness as numbers and fast_food as boolean can be used. Set to empty to disable this. This is evaluated with EvalEx https://ezylang.github.io/EvalEx/concepts/parsing_evaluation.html.")
-    public static String foodStackReductionFormula = "ROUND(MAX(64 / hunger * 0.4, 1), 0)";
+    public static String foodStackReductionFormula = "ROUND(MAX(64 / MAX(hunger, 1) * 0.4, 1), 0)";
     @Config(min = 0.01d, max = 64d)
     @Label(name = "Item Stack Multiplier", description = "Items max stack sizes (excluding blocks) will be multiplied by this value. Foods will be overridden by 'Food Stack Reduction' or 'Food Stack Multiplier' if are active. Setting to 1 will disable this feature.")
     public static Double itemStackMultiplier = 1d;